pyproject.toml
[tool.poetry]
name = "supply-chain-detector"
version = "0.1.0"
description = "AI-powered software supply chain attack detector"
authors = ["Mark Christian Anub"]
readme = "README.md"
packages = [
	{ include = "fetcher" },
	{ include = "api" },
	{ include = "detector" },
	{ include = "storage" },
	{ include = "workers" },
	{ include = "ml" }
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.11.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
tqdm = "^4.67.1"
sentence-transformers = "^3.4.1"
faiss-cpu = "^1.10.0"
networkx = "^3.4"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
celery = "^5.4.0"
redis = "^5.2.0"
streamlit = "^1.43.0"
sqlalchemy = "^2.0.39"
psycopg = {version = "^3.2.9", extras = ["binary"]}
xgboost = "^2.1.4"
scikit-learn = "^1.6.1"
alembic = "^1.13.0"
seaborn = "^0.13"
umap-learn = "^0.5"
scipy = "^1.14"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.11.2"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["notebooks/*.ipynb"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
